Add Player object + player naming test
